licitometro/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
pytest==8.0.0
pytest-asyncio==0.23.5
python-multipart==0.0.6
python-dotenv==1.0.0
minio==7.2.0
redis==5.0.1
elasticsearch[async]==8.11.1
httpx==0.26.0
pytest-cov==4.1.0
celery==5.3.6
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
aiohttp==3.9.3
